backend/requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
Werkzeug==2.3.0
tensorflow==2.12.0
opencv-python-headless==4.7.0.72
numpy==1.24.3
Pillow==9.5.0
